+++ lib/phoenix_component.ex
@@ -751,7 +751,7 @@ defmodule Phoenix.Component do
       caller: __CALLER__,
       indentation: meta[:indentation] || 0,
       source: expr,
-      tag_handler: Phoenix.LiveView.Tokenizer.HTML
+      tag_handler: Phoenix.LiveView.HTMLEngine
     ]
 
     EEx.compile_string(expr, options)
